Deleted handlers.go and wmiau.go
